+++ .github/workflows/benchmark.yml
@@ -27,7 +27,7 @@ jobs:
       - name: Build code
         run: yarn transpile && yarn build:node-release
       - name: 'Run benchmark'
-        run: yarn benchmark
+        run: yarn benchmark:node
 
       - name: Download previous benchmark data
         uses: actions/cache@v1
