+++ bench/setup.js
@@ -1,3 +1,3 @@
-const { default: SourceMap, init } = process.env.BACKEND === 'wasm' ? require('../dist/wasm-node') : require('..');
+const { default: SourceMap, init } = process.env.BACKEND === 'wasm' ? require('../dist/wasm') : require('..');
 
 module.exports = { SourceMap, init };
