+++ src/node.js
@@ -30,35 +30,6 @@ export default class NodeSourceMap extends SourceMap {
     return this;
   }
 
-  addIndexedMappings(
-    mappings: Array<IndexedMapping<string>>,
-    lineOffset?: number = 0,
-    columnOffset?: number = 0
-  ): SourceMap {
-    let mappingBuffer = this._indexedMappingsToInt32Array(mappings, lineOffset, columnOffset);
-    this.sourceMapInstance.addIndexedMappings(mappingBuffer);
-    return this;
-  }
-
-  fromBuffer(buffer: Buffer): this {
-    this.sourceMapInstance.fromBuffer(buffer);
-    return this;
-  }
-
-  toBuffer(): Buffer {
-    return this.sourceMapInstance.toBuffer();
-  }
-
-  findClosestMapping(line: number, column: number): ?IndexedMapping<string> {
-    let mapping = this.sourceMapInstance.findClosestMapping(line - 1, column);
-    if (mapping) {
-      let v = this.indexedMappingToStringMapping(mapping);
-      return v;
-    } else {
-      return null;
-    }
-  }
-
   addSourceMap(sourcemap: SourceMap, lineOffset: number = 0, columnOffset: number = 0): SourceMap {
     if (!(sourcemap.sourceMapInstance instanceof bindings.SourceMap)) {
       throw new Error('The sourcemap provided to addSourceMap is not a valid sourcemap instance');
